Fix address reading and user int type

+++ src/common/command/getaccount.h
@@ -13,7 +13,7 @@
 class GetAccount : public BlockCommand {
   public:
     GetAccount();
-    GetAccount(uint16_t abank, uint32_t auser, uint16_t bbank, uint16_t buser, uint32_t time);
+    GetAccount(uint16_t abank, uint32_t auser, uint16_t bbank, uint32_t buser, uint32_t time);
 
 
     //IBlock interface
